Extracted remaining helper function into utilities code and header file.
